requirements.txt
flask==2.1.2
requests==2.27.1
gevent==21.12.0
joblib==1.1.0
loguru==0.6.0
PyYAML==6.0.0